feat: user entity version

+++ src/main/java/org/example/owner/OwnersController.java
@@ -7,11 +7,11 @@ import org.example.exceptions.IncorrectArgumentsException;
 import org.example.infrastructure.UserIdentitySecurityChecker;
 import org.example.owner.dto.OwnerIdResponse;
 import org.example.owner.dto.OwnerInfoResponse;
-import org.example.owner.dto.UserLoginDto;
-import org.example.valueObjects.Role;
+import org.example.user.UserLoginDto;
 import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -29,30 +29,12 @@ public class OwnersController {
     private final ManagingCatsUsecases managingCatsUsecases;
     private final AuthenticationManager authenticationManager;
     private final UserIdentitySecurityChecker securityChecker;
-    /*@PostMapping("/owner/token")
-    public ResponseEntity<JwtAuthToken> getToken(@RequestBody UserLoginInfo userLoginInfo) {
-        var token = service.getToken(userLoginInfo.getUsername(), userLoginInfo.getPassword());
-        if (token == null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        return new ResponseEntity<>(new JwtAuthToken(token), HttpStatus.OK);
-    }*/
-
-    @PostMapping("/owner/login")
-    public ResponseEntity<String> loginOwner(@RequestBody UserLoginDto loginDto){
-        Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        loginDto.getUsername(),
-                        loginDto.getPassword()));
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        return new ResponseEntity<>("Successfully logged in", HttpStatus.OK);
-    }
-
-    @PreAuthorize("#ownerId == authentication.principal.id" + "|| hasRole('ROLE_ADMIN')")
     @GetMapping("/owner/{ownerId}")
     public ResponseEntity<OwnerInfoResponse> getOwnerById(@PathVariable long ownerId) {
-        //if (!securityChecker.isAdmin() && !UserIdentitySecurityChecker.checkUserIdentity(ownerId)) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        if (!securityChecker.isAdmin() && !securityChecker.checkOwnerIdentity(ownerId)) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         var returnedOwner = service.getOwnerById(ownerId);
         if (returnedOwner == null) return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        return new ResponseEntity<>(new OwnerInfoResponse(returnedOwner.getId(), returnedOwner.getBirthday(), returnedOwner.getCats().stream().map(Cat::getId).toList(), returnedOwner.getUsername(), returnedOwner.getAuthorities()),
+        return new ResponseEntity<>(new OwnerInfoResponse(returnedOwner.getId(), returnedOwner.getBirthday(), returnedOwner.getCats().stream().map(Cat::getId).toList()),
                 HttpStatus.OK);
     }
     @GetMapping("/owners")
@@ -61,7 +43,7 @@ public class OwnersController {
         var criteria = new FindCriteria(birthday);
         var returnedOwner = service.getOwnerByCriteria(criteria);
         if (returnedOwner == null) return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        return new ResponseEntity<>(returnedOwner.stream().map(x -> new OwnerInfoResponse(x.getId(), x.getBirthday(), x.getCats().stream().map(Cat::getId).toList(), x.getUsername(), x.getAuthorities())).toList(), HttpStatus.OK);
+        return new ResponseEntity<>(returnedOwner.stream().map(x -> new OwnerInfoResponse(x.getId(), x.getBirthday(), x.getCats().stream().map(Cat::getId).toList())).toList(), HttpStatus.OK);
     }
     @PostMapping("/owner")
     public ResponseEntity<OwnerIdResponse> save(@RequestBody OwnerInfoDto dto) {
@@ -75,10 +57,9 @@ public class OwnersController {
                     HttpStatus.BAD_REQUEST);
         }
     }
-    @PreAuthorize("#ownerId == authentication.principal.id" + "|| hasRole('ROLE_ADMIN')")
     @PutMapping("/owner/{ownerId}")
     public ResponseEntity<OwnerIdResponse> update(@RequestBody OwnerInfoDto dto, @PathVariable long ownerId) {
-        //if (!securityChecker.isAdmin() && !UserIdentitySecurityChecker.checkUserIdentity(ownerId)) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        if (!securityChecker.isAdmin() && !securityChecker.checkOwnerIdentity(ownerId)) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         Owner returnedOwner;
         try {
             returnedOwner = service.update(dto);
@@ -89,10 +70,9 @@ public class OwnersController {
         return new ResponseEntity<>(new OwnerIdResponse(returnedOwner.getId()),
                 HttpStatus.OK);
     }
-    @PreAuthorize("#ownerId == authentication.principal.id" + "|| hasRole('ROLE_ADMIN')")
     @DeleteMapping("/owner/{ownerId}")
-    //if (!securityChecker.isAdmin() && !UserIdentitySecurityChecker.checkUserIdentity(ownerId)) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     public void delete(@PathVariable long ownerId) {
+        if (!securityChecker.isAdmin() && !securityChecker.checkOwnerIdentity(ownerId)) throw new AccessDeniedException("Access denied");
         service.delete(ownerId);
     }
 }
